created wikipedia_router.js in helper folder and added to MongoClient
